+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
  - Fix Wine crash on newer Steam Runtime installations due to renamed runtime executable
  - Fix graphical Wine applications crashing on Wayland
+ - Fix Protontricks crash caused by Steam shortcuts created by 3rd party applications such as Lutris
 
 ## [1.8.1] - 2022-03-20
 ### Added